Refactor vue stuff
